Fix tests

+++ pkg/manager/uninstall.go
@@ -6,6 +6,8 @@ import (
 	"path/filepath"
 
 	"github.com/hashicorp/go-version"
+
+	"github.com/NoizeMe/go-man/pkg/utils"
 )
 
 // UninstallAll is a function that removes all current installations of the Go SDK.
@@ -39,6 +41,9 @@ func (m *GoManager) Uninstall(versionNumber *version.Version) error {
 	uninstallTask := m.task.Step()
 	uninstallTask.Printf("Deleting SDK: %s", versionDirectory)
 
+	if !utils.PathExists(versionDirectory) {
+		return fmt.Errorf("no directory %s to uninstall from", versionDirectory)
+	}
 	if err := deleteVersionDirectory(versionDirectory); err != nil {
 		return err
 	}
